Moving ArticleList to a separate folder
